Changing to long int some variables were possibly needed. This should be checked better in the future running on single MPI multiple thread.

+++ src/linsolve.h
@@ -36,7 +36,7 @@
                             gmres_float_struct *p, level_struct *l, struct Thread *threading );
   
   void fgmres_MP_struct_init( gmres_MP_struct *p );
-  void fgmres_MP_struct_alloc( int m, int n, int vl, double tol, const int prec_kind,
+  void fgmres_MP_struct_alloc( int m, int n, long int vl, double tol, const int prec_kind,
                                void (*precond)(), gmres_MP_struct *p, level_struct* l );
   void fgmres_MP_struct_free( gmres_MP_struct *p );
